[TML] rename concept NetworkBettinaVar to NetworkBettina
