+++ TreeModelLib/GrowthAndDeathDynamics/NetworkBettina/NetworkBettina.py
@@ -4,6 +4,7 @@
 @date: 2021-Today
 @author: marie-christin.wimmler@tu-dresden.de
 """
+
 from TreeModelLib.GrowthAndDeathDynamics import GrowthAndDeathDynamics
 from TreeModelLib.GrowthAndDeathDynamics.SimpleBettina import SimpleBettina
 import numpy as np
@@ -27,10 +28,11 @@ class NetworkBettina(SimpleBettina):
         self.potential_partner = network['potential_partner']
         self.psi_osmo = network["psi_osmo"]
 
-        self.r_gr_min = network['r_gr_min']
-        self.r_gr_rgf = network['r_gr_rgf']
-        self.l_gr_rgf = network['l_gr_rgf']
-        self.weight_gr = network['weight_gr']
+        if self.variant == "V2":
+            self.r_gr_min = network['r_gr_min']
+            self.r_gr_rgf = network['r_gr_rgf']
+            self.l_gr_rgf = network['l_gr_rgf']
+            self.weight_gr = network['weight_gr']
 
         self.name = str(tree.group_name) + str(tree.tree_id)
         # Simple bettina tree progress
@@ -41,10 +43,12 @@ class NetworkBettina(SimpleBettina):
         network['potential_partner'] = self.potential_partner
         network['partner'] = self.partner
 
-        network['r_gr_min'] = self.r_gr_min
-        network['r_gr_rgf'] = self.r_gr_rgf
-        network['l_gr_rgf'] = self.l_gr_rgf
-        network['weight_gr'] = self.weight_gr
+        if self.variant == "V2":
+            network['r_gr_min'] = self.r_gr_min
+            network['r_gr_rgf'] = self.r_gr_rgf
+            network['l_gr_rgf'] = self.l_gr_rgf
+            network['weight_gr'] = self.weight_gr   # only required for csv
+            # output
 
         tree.setNetwork(network)
         if self.survive == 1:
@@ -52,30 +56,68 @@ class NetworkBettina(SimpleBettina):
         else:
             tree.setSurvival(0)
 
+    def treeGrowthWeights(self):
+        if self.variant == 'V2':
+            self.treeGrowthWeightsV2()
+        else:
+            SimpleBettina.treeGrowthWeights(self)
+
+    def growthResources(self):
+        SimpleBettina.growthResources(self)
+        if self.variant == "V0":
+            self.rootGraftFormationV0()
+        if self.variant == "V1":
+            self.growthResourcesV1()
+
+
+
     ## This functions calculates the growths weights for distributing
     # biomass increment to the geometric (allometric) tree measures as
     # defined in SimpleBettina. In addition, the function calls the root
     # graft formation function, if the tree is currently in the process of
     # root graft formation.
-    def treeGrowthWeights(self):
+    def treeGrowthWeightsV2(self):
         # Simple bettina get growth weigths
         SimpleBettina.treeGrowthWeights(self)
 
         # If self.r_gr_min exists the tree is currently in the process of rgf
         if self.r_gr_min:
-            self.rootGraftFormation()
+            self.rootGraftFormationV2()
         else:
             self.weight_gr = 0
 
+    ## This function calculates the available resources and the biomass
+    # increment. In addition, this function reduces the available resources
+    # if trees are in the grafting process and calls the root graft
+    # formation manager
+    def growthResourcesV1(self):
+        # Simple bettina get growth resources
+        #SimpleBettina.growthResources(self)
+        self.rootGraftFormationV1()
+
+    def rootGraftFormationV0(self):
+        if self.rgf != -1:
+            self.rgf = -1
+            self.partner.append(self.potential_partner)
+            self.potential_partner = []
+
+    ## This function reduces growth during the process of root graft formation.
+    # It is assumed that the process will take 2 years (this is subject to
+    # change).
+    def rootGraftFormationV1(self):
+        if self.rgf != -1:
+            self.grow = self.grow * (1 - self.f_growth)
+            self.rgf = self.rgf + 1
+            if round(self.rgf * self.time / 3600 / 24 / 365, 3) >= 2:
+                self.rgf = -1
+                self.partner.append(self.potential_partner)
+                self.potential_partner = []
 
     ## This function simulated root graft formation and thereby reduces girth
     # growth during this process.
-    def rootGraftFormation(self):
+    def rootGraftFormationV2(self):
         # check if rgf is finished
         if self.r_gr_rgf >= self.r_gr_min[0]:
-            print('RGF is finished after ' + str(self.rgf) + ' time '
-                 'steps. Years: ' + str(self.rgf * self.time / 3600 / 24 /
-                                        365))
             # Append partner
             self.partner.append(self.potential_partner)
             # Reset rgf parameters
@@ -86,22 +128,23 @@ class NetworkBettina(SimpleBettina):
             self.r_gr_min = []
         else:
             # if rgf is not finished, the grafted roots must grow
-            self.weight_gr = self.weight_girthgrowth * self.f_wgr
+            self.weight_gr = self.weight_girthgrowth * self.f_growth
             inc_r_gr = (self.weight_gr * self.grow /
                         (2 * np.pi * self.l_gr_rgf * self.r_gr_rgf))
             self.r_gr_rgf += inc_r_gr
             # reduce girth growth factor
             self.weight_girthgrowth = self.weight_girthgrowth * (1 -
-                                                                 self.f_wgr)
+                                                                 self.f_growth)
             self.rgf += 1
 
     def getInputParameters(self, args):
-        missing_tags = ["type", "f_wgr"]
+        missing_tags = ["type", "variant", "f_growth"]
         for arg in args.iterdescendants():
             tag = arg.tag
-            print('tag ' + str(tag))
-            if tag == "f_wgr":
-                self.f_wgr = float(args.find("f_wgr").text)
+            if tag == "variant":
+                self.variant = args.find("variant").text
+            if tag == "f_growth":
+                self.f_growth = float(args.find("f_growth").text)
             try:
                 missing_tags.remove(tag)
             except ValueError:
@@ -116,4 +159,9 @@ class NetworkBettina(SimpleBettina):
                 "Tag(s) " + string +
                 "are not given for growth and death initialisation in "
                 "project file."
+            )
+        if self.variant not in ["V0", "V1", "V2"]:
+            raise KeyError(
+                "NetworkBettina variant " + self.variant +
+                " is not defined. Existing variants are 'V0', 'V1' and 'V2'."
             )
